S11P12D109-248 feat: 공연, 극장, 티켓 컨트롤러 추가

+++ Colorpl-Backend/src/main/java/com/colorpl/show/domain/detail/ShowDetailRepository.java
@@ -3,7 +3,7 @@ package com.colorpl.show.domain.detail;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShowDetailRepository extends JpaRepository<ShowDetail, Integer> {
+public interface ShowDetailRepository extends JpaRepository<ShowDetail, Long> {
 
     Optional<ShowDetail> findByApiId(String apiId);
 }
